+++ src/main/java/umc/duckmelang/domain/auth/service/AuthService.java
@@ -8,14 +8,14 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import umc.duckmelang.domain.auth.client.KakaoApiClient;
 import umc.duckmelang.domain.auth.dto.response.LoginResponse;
 import umc.duckmelang.domain.auth.service.strategy.SocialLoginStrategy;
 import umc.duckmelang.domain.member.domain.Member;
+import umc.duckmelang.domain.member.domain.enums.LoginType;
 import umc.duckmelang.domain.member.repository.MemberRepository;
 import umc.duckmelang.global.apipayload.exception.MemberException;
 import umc.duckmelang.global.apipayload.exception.TokenException;
-import umc.duckmelang.domain.auth.redis.RefreshTokenServiceImpl;
+import umc.duckmelang.domain.auth.refreshToken.RefreshTokenServiceImpl;
 import umc.duckmelang.domain.auth.jwt.JwtTokenProvider;
 import umc.duckmelang.domain.auth.user.CustomUserDetails;
 import umc.duckmelang.global.apipayload.code.status.ErrorStatus;
@@ -30,7 +30,7 @@ public class AuthService {
     private final JwtTokenProvider jwtTokenProvider;
     private final RefreshTokenServiceImpl refreshTokenService;
     private final MemberRepository memberRepository;
-    private final Map<String, SocialLoginStrategy> strategyMap;
+    private final Map<String, SocialLoginStrategy> loginStrategyMap;
 
     // 자체 로그인
     @Transactional
@@ -54,6 +54,16 @@ public class AuthService {
         } catch (BadCredentialsException e) {
             throw new AuthException(ErrorStatus.AUTH_INVALID_CREDENTIALS);
         }
+    }
+
+    // 소셜 로그인 처리
+    @Transactional
+    public LoginResponse socialLogin(LoginType loginType, String accessToken){
+        SocialLoginStrategy strategy = loginStrategyMap.get(loginType.name());
+        if(strategy == null){
+            throw new AuthException(ErrorStatus.AUTH_INVALID_CREDENTIALS);
+        }
+        return strategy.login(accessToken);
     }
 
     // 토큰 재발급
